reworked structure of the project
