+++ CMakeLists.txt
@@ -18,18 +18,18 @@ add_custom_command(
     COMMAND cwd=$(pwd) && cd ${LIBPCAP} && pwd && ./configure --enable-dbus=no --enable-bluetooth=no --enable-usb=no && make && cd ${cwd}
 )
 
-add_library (cpppcap STATIC ${LIBPCAP_STATIC})
+add_library (cpppcap STATIC include/kp_analysis_from_dump.hpp sources/kp_analysis_from_dump.cpp ${LIBPCAP_STATIC})
 add_library (pcap STATIC IMPORTED GLOBAL)
 set_target_properties(pcap PROPERTIES IMPORTED_LOCATION ${LIBPCAP_STATIC})
 link_directories(${LIBPCAP})
 include_directories(${LIBPCAP})
 
-add_executable(analysis_from_dump
-  ${CMAKE_CURRENT_SOURCE_DIR}/sources/kp_analysis_from_dump.cpp
+add_executable(test
+  ${CMAKE_CURRENT_SOURCE_DIR}/tests/simple.cpp
 )
 
-target_link_libraries(analysis_from_dump cpppcap pcap)
-
+target_link_libraries(test cpppcap)
+target_link_libraries(test pcap)
 
 #include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include)
