+++ tests/simple.cpp
@@ -1,0 +1,6 @@
+﻿#include <kp_analysis_from_dump.hpp>
+
+int main() {
+	main_f();
+	return 0;
+}
